+++ src/main/java/roomescape/auth/constants/AuthConstants.java
@@ -1,4 +1,4 @@
-package roomescape.auth;
+package roomescape.auth.constants;
 
 public class AuthConstants {
     public static final String AUTH_TOKEN_COOKIE = "token";
