+++ src/main/java/roomescape/common/config/WebConfig.java
@@ -1,12 +1,12 @@
-package roomescape.auth.config;
+package roomescape.common.config;
 
 import java.util.List;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.auth.AdminInterceptor;
-import roomescape.auth.LoginMemberArgumentResolver;
+import roomescape.common.interceptor.AdminInterceptor;
+import roomescape.common.resolver.LoginMemberArgumentResolver;
 
 @Configuration
 public class WebConfig implements WebMvcConfigurer {
