+++ src/main/java/roomescape/common/interceptor/AdminInterceptor.java
@@ -1,10 +1,12 @@
-package roomescape.auth;
+package roomescape.common.interceptor;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.HandlerInterceptor;
-import roomescape.member.Member;
+import roomescape.auth.service.AuthService;
+import roomescape.auth.util.CookieManager;
+import roomescape.member.dto.Member;
 
 @Component
 public class AdminInterceptor implements HandlerInterceptor {
