+++ src/main/java/roomescape/member/dao/MemberDao.java
@@ -1,9 +1,10 @@
-package roomescape.member;
+package roomescape.member.dao;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
+import roomescape.member.dto.Member;
 
 @Repository
 public class MemberDao {
