+++ src/main/java/roomescape/reservation/controller/ReservationController.java
@@ -1,6 +1,5 @@
-package roomescape.reservation;
+package roomescape.reservation.controller;
 
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -11,7 +10,10 @@ import org.springframework.web.bind.annotation.RestController;
 
 import java.net.URI;
 import java.util.List;
-import roomescape.member.LoginMember;
+import roomescape.auth.dto.LoginMember;
+import roomescape.reservation.dto.ReservationRequest;
+import roomescape.reservation.dto.ReservationResponse;
+import roomescape.reservation.service.ReservationService;
 
 @RestController
 public class ReservationController {
