+++ src/main/java/roomescape/reservation/dao/ReservationDao.java
@@ -1,9 +1,11 @@
-package roomescape.reservation;
+package roomescape.reservation.dao;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
+import roomescape.reservation.dto.Reservation;
+import roomescape.reservation.dto.ReservationRequest;
 import roomescape.theme.Theme;
 import roomescape.time.Time;
